refactor: address ninth-round PR #608 review
Apply normalizeBrailleDisplay at SettingsService construction, right
after deepMerge but before any consumer reads the settings. Previously
the repair only ran inside the Settings component's useState
initializer, so BrailleService and other consumers could observe a
stale brailleDisplayKind / brailleDisplayPresetId between page load
and the first time the user opened the settings dialog. Now the
service hands out a coherent state from the moment it constructs.

Move SINGLE_LINE_BRAILLE_PRESETS and MULTI_LINE_BRAILLE_PRESETS from
src/type/settings.ts into src/util/braillePreset.ts. The arrays are
runtime data, not type declarations, and they sit best alongside the
helper functions that consume them. The types file now keeps the
shape declarations (BrailleDisplayPreset, NonEmptyBraillePresets,
BraillePresetSelection, BrailleDisplayKind) plus the small
configuration constants. Imports in Settings.tsx and the test file
updated accordingly.

Use parseManualBrailleInput on the manual size/lines onBlur handlers
too, so a user who clears the field and tabs away no longer has the
empty value silently coerced to 1 (Number('') === 0, clamp(0) === 1).
Empty-on-blur now skips the update for parity with onChange,
preserving whatever value was last stored.
